Move decoder and encoder to codec module
