+++ wsidicom/codec/__init__.py
@@ -1,0 +1,15 @@
+from wsidicom.codec.encoder import Encoder
+from wsidicom.codec.settings import (
+    Channels,
+    JpegSettings,
+    JpegLsNearLosslessSettings,
+    JpegLosslessSettings,
+    Jpeg2kLosslessSettings,
+    Jpeg2kSettings,
+    JpegLsLosslessSettings,
+    NumpySettings,
+    RleSettings,
+    Settings,
+    Subsampling,
+)
+from wsidicom.codec.decoder import Decoder
